requirements.txt
httpx
beautifulsoup4

discord.py
google-cloud-texttospeech
PyNaCl

aiosqlite

python-dotenv